Rename meta.py to package.py to reflect that it is package level configuration.
